Prefer direct Ascend C runtime in tests and CI
