+++ torch_custom/fla_npu/test/test_fwd_o.py
@@ -1,3 +1,13 @@
+# -----------------------------------------------------------------------------------------------------------
+# Copyright (c) 2026 Tianjin University, Ltd.
+# This program is free software, you can redistribute it and/or modify it under the terms and conditions of
+# CANN Open Software License Agreement Version 2.0 (the "License").
+# Please refer to the License for details. You may not use this file except in compliance with the License.
+# THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED,
+# INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
+# See LICENSE in the root of the software repository for the full text of the License.
+# -----------------------------------------------------------------------------------------------------------
+
 import os
 import sys
 import logging
@@ -5,14 +15,12 @@ import numpy as np
 import random
 import ml_dtypes
 import torch
-import torch_npu
 from ml_dtypes import bfloat16
 from dataclasses import dataclass
 import math
 # import custom_ops
-import fla_npu
+from fla_npu.ops import ascendc as ascendc_ops
 
-fla_npu.load_legacy_torch_ops()
 
 torch.npu.config.allow_internal_format = False
 torch.npu.set_compile_mode(jit_compile=False)
@@ -313,7 +321,7 @@ if __name__ == "__main__":
     torch.npu.synchronize()
     print("step 6: before custom op")
 
-    result = torch.ops.npu.npu_chunk_fwd_o(
+    result = ascendc_ops.npu_chunk_fwd_o(
         q,
         k,
         v,
