+++ torch_custom/fla_npu/test/test_npu_causal_conv1d.py
@@ -1,13 +1,21 @@
+# -----------------------------------------------------------------------------------------------------------
+# Copyright (c) 2026 Tianjin University, Ltd.
+# This program is free software, you can redistribute it and/or modify it under the terms and conditions of
+# CANN Open Software License Agreement Version 2.0 (the "License").
+# Please refer to the License for details. You may not use this file except in compliance with the License.
+# THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED,
+# INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
+# See LICENSE in the root of the software repository for the full text of the License.
+# -----------------------------------------------------------------------------------------------------------
+
 import unittest
 
 import os
 import torch
-import torch_npu
 import torch.nn.functional as F
 
-import fla_npu
+from fla_npu.ops import ascendc as ascendc_ops
 
-fla_npu.load_legacy_torch_ops()
 
 torch.npu.set_device(int(os.environ.get("TEST_DEVICE_ID", 0)))
 
@@ -178,8 +186,7 @@ class TestCausalConv1d(unittest.TestCase):
     atol = 5e-2
 
     def call_op(self, **kwargs):
-        self.assertTrue(hasattr(torch.ops.npu, "npu_causal_conv1d"))
-        return torch.ops.npu.npu_causal_conv1d(**kwargs)
+        return ascendc_ops.npu_causal_conv1d(**kwargs)
 
     def assertTensorClose(self, actual: torch.Tensor, expected: torch.Tensor, *, rtol=None, atol=None):
         rtol = self.rtol if rtol is None else rtol
