+++ torch_custom/fla_npu/test/test_npu_recompute_w_u_fwd.py
@@ -1,14 +1,22 @@
+# -----------------------------------------------------------------------------------------------------------
+# Copyright (c) 2026 Tianjin University, Ltd.
+# This program is free software, you can redistribute it and/or modify it under the terms and conditions of
+# CANN Open Software License Agreement Version 2.0 (the "License").
+# Please refer to the License for details. You may not use this file except in compliance with the License.
+# THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED,
+# INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
+# See LICENSE in the root of the software repository for the full text of the License.
+# -----------------------------------------------------------------------------------------------------------
+
 import torch
-import torch_npu
 import os
 from typing import Optional
 import pickle
 import math
 import ct
 import random
-import fla_npu
+from fla_npu.ops import ascendc as ascendc_ops
 
-fla_npu.load_legacy_torch_ops()
 torch.npu.set_device(int(os.environ.get("TEST_DEVICE_ID", 0)))
 
 def get_bos_eos(idx, T, chunk_size, cu_seqlens, chunk_indices):
@@ -279,7 +287,7 @@ def test_recompute_wu_fwd(
 
     # 将张量移到 NPU 并调用反向算子
     if chunk_indices != None:
-        w, u = torch.ops.npu.npu_recompute_w_u_fwd(
+        w, u = ascendc_ops.npu_recompute_w_u_fwd(
             k.npu(),
             v.npu(),
             beta.npu(),
@@ -291,7 +299,7 @@ def test_recompute_wu_fwd(
             chunk_indices=chunk_indices
         )
     else:
-        w, u = torch.ops.npu.npu_recompute_w_u_fwd(
+        w, u = ascendc_ops.npu_recompute_w_u_fwd(
             k.npu(),
             v.npu(),
             beta.npu(),
